requirements.txt
Flask==3.1.2
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
python-dotenv==1.2.1
SQLAlchemy==2.0.45
WTForms==3.2.1
